+2 альтернативна задачка
